requirements.txt
streamlit==1.20.0
transformers==4.25.1
torch==1.13.1
requests==2.28.1
python-dotenv==1.0.0
